refactor(cli): split controller and entity files
