+++ cmd/fuego/commands/controller_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestCreateController(t *testing.T) {
-	res, err := createController("books")
+	res, err := createControllerFile("books", "controller.go", "booksController.go")
 	require.NoError(t, err)
 	require.Contains(t, res, "package controller")
 	require.Contains(t, res, `fuego.Get(booksGroup, "/{id}", rs.getBooks)`)
